+++ hari_client/upload/hari_uploader.py
@@ -1,5 +1,7 @@
 import copy
+import re
 import typing
+import urllib.parse
 import uuid
 
 import pydantic
@@ -170,6 +172,43 @@ class HARIMedia(models.BulkMediaCreate):
             )
         return v
 
+    @pydantic.field_validator("media_url")
+    @classmethod
+    def media_url_is_valid(cls, v: str) -> str:
+        # In case we wanted to allow simple relative paths, too:
+        # if not v.startswith(("s3://", "http://", "https://")):
+        #     if re.match(r"^[^\s/]+(\.[a-zA-Z0-9]+)?$", v) or "/" in v:
+        #         return v
+        #     raise ValueError("Invalid URL or path format")
+
+        parsed_url = urllib.parse.urlparse(v)
+
+        if parsed_url.scheme not in ("https", "s3"):
+            raise ValueError("URL must start with https:// or s3://")
+
+        if parsed_url.scheme == "s3":
+            if not parsed_url.netloc:
+                raise ValueError("Invalid S3 URL format")
+        else:
+            # Azure Blob Storage HTTPS URL pattern
+            azure_pattern = re.compile(
+                r"^https://[a-z0-9]{3,24}\.blob\.core\.windows\.net/[^\s]+$",
+                re.IGNORECASE,
+            )
+
+            # AWS S3 HTTPS URL pattern (must include region)
+            aws_pattern = re.compile(
+                r"^https://([a-z0-9.-]+\.)?s3[.-][a-z0-9-]+\.amazonaws\.com/[^\s]+$",
+                re.IGNORECASE,
+            )
+
+            if not (azure_pattern.match(v) or aws_pattern.match(v)):
+                raise ValueError(
+                    "Invalid HTTPS URL format for S3 or Azure Blob Storage"
+                )
+
+        return v
+
 
 class HARIUploadResults(pydantic.BaseModel):
     medias: models.BulkResponse
@@ -205,6 +244,7 @@ class HARIUploader:
         # TODO: this should be a dict[str, uuid.UUID] as soon as the api models are updated
         self._object_category_subsets: dict[str, str] = {}
         self._unique_attribute_ids: set[str] = set()
+        self._with_media_file_upload: bool = True
 
     # TODO: add_media shouldn't do validation logic, because that expects that a specific order of operation is necessary,
     # specifically that means that media_objects and attributes have to be added to media before the media is added to the uploader.
@@ -415,6 +455,29 @@ class HARIUploader:
                 all_attributes.extend(media_object.attributes)
         validation.validate_attributes(all_attributes)
 
+    def _prepare_media_upload(self) -> None:
+        """Checks whether medias have to be uploaded or not, based on their file_path and media_url being set."""
+        with_media_file_upload_cnt = 0
+        with_media_url_cnt = 0
+        for media in self._medias:
+            if media.file_path:
+                with_media_file_upload_cnt += 1
+            elif media.media_url:
+                with_media_url_cnt += 1
+
+        if with_media_file_upload_cnt == len(self._medias):
+            self._with_media_file_upload = True
+        elif (
+            with_media_url_cnt == len(self._medias) and with_media_file_upload_cnt == 0
+        ):
+            self._with_media_file_upload = False
+        else:
+            raise HARIMediaUploadError(
+                "Found mixed file_path and media_url usage in HARIMedia. Use only one consistently."
+                + " If you're using an external media source, don't set the file_path of HARIMedia and only set the media_url."
+                + " If you're not using an external media source, don't set the media_url of HARIMedia and only set the file_path."
+            )
+
     def upload(
         self,
     ) -> HARIUploadResults | None:
@@ -451,6 +514,7 @@ class HARIUploader:
             )
 
         self.validate_all_attributes()
+        self._prepare_media_upload()
 
         self._handle_object_categories()
 
@@ -507,7 +571,9 @@ class HARIUploader:
 
         # upload media batch
         media_upload_response = self.client.create_medias(
-            dataset_id=self.dataset_id, medias=medias_to_upload
+            dataset_id=self.dataset_id,
+            medias=medias_to_upload,
+            with_media_file_upload=self._with_media_file_upload,
         )
         self._media_upload_progress.update(len(medias_to_upload))