+++ CHANGELOG.md
@@ -9,6 +9,16 @@
 - added support for defining external media sources when creating a dataset
   - new field `external_media_source` in the `create_dataset` method [PR#73](https://github.com/quality-match/hari-client/pull/73)
 - added new endpoint `get_external_media_source` [PR#73](https://github.com/quality-match/hari-client/pull/73)
+- added new arg to client method `create_medias` [PR#74](https://github.com/quality-match/hari-client/pull/74)
+  - `with_media_file_upload` (default: `True`). Set this to `False` if you want to skip the upload of media files.
+  - the method will then only send the `BulkMediaCreate` instance to HARI.
+- added format validations for `HARIMedia` field `media_url` [PR#74](https://github.com/quality-match/hari-client/pull/74)
+  - the field was previously set automatically through the process of uploading media files.
+  - in order to work with a dataset that uses an external media source, you have to set the field explicitly.
+- updated usage of HARIUploader utility [PR#74](https://github.com/quality-match/hari-client/pull/74)
+  - when using an external media source, make sure to only set the `media_url` field of a `HARIMedia` and don't set the `file_path`.
+  - if all medias have a set `media_url` field, then media files will not be uploaded to HARI, but they're only setup with the api.
+  - if all medias have a set `file_path` field, then the uploader utility behaves as before and will upload media files to HARI.
 
 ## [3.2.0] - 25-02-2025